+++ assets/js/common/DottedPagination/index.js
