+++ assets/js/pages/ClusterDetails/AscsErsClusterDetails.jsx
@@ -11,7 +11,7 @@ import BackButton from '@common/BackButton';
 import Table from '@common/Table';
 import ListView from '@common/ListView';
 import ProviderLabel from '@components/ProviderLabel';
-import DottedPagination from '@components/DottedPagination';
+import DottedPagination from '@common/DottedPagination';
 import ClusterNodeLink from '@pages/ClusterDetails/ClusterNodeLink';
 import SapSystemLink from '@components/SapSystemLink';
 import Tooltip from '@common/Tooltip';
